requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
appwrite==2.0.2
asgiref==3.6.0
async-timeout==4.0.2
atomicwrites==1.4.0
attrs==23.1.0
backports.weakref==1.0.post1
brotlipy==0.7.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
chromedriver-autoinstaller==0.6.2
ci-info==0.3.0
click==8.1.7
configobj==5.0.8
configparser==5.3.0
constantly==15.1.0
cryptography==41.0.4
cssselect==1.1.0
dataclasses-json==0.6.1
deprecation==2.1.0
Django==4.2.1
django-cors-headers==4.3.0
django-environ==0.11.2
django-ratelimit==4.1.0
django-tinymce==3.6.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
dnspython==2.4.0
docstring-to-markdown==0.13
docx==0.2.4
docx2python==2.6.3
docx2txt==0.8
et-xmlfile==1.1.0
etelemetry==0.3.0
exceptiongroup==1.1.2
faiss-cpu==1.7.4
filelock==3.12.4
fitz==0.0.1.dev2
fonttools==4.25.0
frozenlist==1.3.3
future==0.18.3
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.60.0
greenlet==3.0.0
h11==0.14.0
HeapDict==1.0.1
httpcore==0.17.3
httplib2==0.22.0
idna==3.4
img2pdf==0.4.4
inflection==0.5.1
iniconfig==2.0.0
isodate==0.6.1
jedi==0.18.2
jsonpatch==1.33
jsonpointer==2.1
langchain==0.0.321
langsmith==0.0.49
llvmlite==0.39.1
looseversion==1.1.2
lxml==4.9.2
marshmallow==3.20.1
mpmath==1.2.1
multidict==6.0.4
munkres==1.1.4
mypy-extensions==0.4.3
networkx==3.2
nibabel==5.1.0
nipype==1.8.6
numpy==1.26.1
oauthlib==3.2.2
openai==0.27.6
openpyxl==3.0.10
outcome==1.2.0
packaging==23.2
pandas==1.5.3
parso==0.8.3
pathlib==1.0.1
patsy==0.5.3
pdfminer.six==20221105
pdfplumber==0.9.0
pep8==1.7.1
pikepdf==7.2.0
Pillow==9.4.0
pluggy==1.3.0
ply==3.11
protobuf==4.24.2
prov==2.0.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycurl==7.45.1
pydantic==2.4.2
pydantic_core==2.10.1
PyDispatcher==2.0.5
pydot==1.4.2
PyJWT==2.8.0
pyls-spyder==0.4.0
pyparsing==3.1.1
PyPDF2==3.0.1
PyQt5-sip==12.11.0
PySocks==1.7.1
pytest==7.1.2
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.7.4
python-pptx==0.6.19
pytz==2023.3.post1
pyxnat==1.5
PyYAML==6.0.1
queuelib==1.5.0
rdflib==6.3.2
regex==2023.10.3
reportlab==4.0.4
requests==2.31.0
requests-oauthlib==1.3.1
rest-framework-simplejwt==0.0.2
rsa==4.9
scipy==1.10.0
selenium==4.14.0
simplejson==3.19.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.22
sqlparse==0.4.4
style==1.1.0
tenacity==8.2.3
tiktoken==0.5.1
tomli==2.0.1
torch==1.12.1
tqdm==4.66.1
traits==6.3.2
trio==0.22.2
trio-websocket==0.10.3
typing-inspect==0.9.0
typing_extensions==4.8.0
ujson==5.8.0
undetected-chromedriver==3.5.0
update==0.0.1
uritemplate==4.1.1
urllib3==1.26.18
validate-email==1.3
Wand==0.6.11
webdriver-manager==4.0.0
webencodings==0.5.1
websockets==11.0.3
whitenoise==6.4.0
wsproto==1.2.0
xlrd==2.0.1
XlsxWriter==3.1.2
yarl==1.9.2
zict==2.1.0
